perf: 修改日志清理
