+++ apps/locale/ja/LC_MESSAGES/django.mo
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4190f4708c72e943e4dbe743c035270b7cfe0bf231383f09cb7b926459c9d74c
-size 135279
+oid sha256:6bbad43c6fd75da933f86aa609c54641cfbf45099564b07e5090d007287af984
+size 135540
