+++ apps/locale/zh/LC_MESSAGES/django.mo
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c33582a24048531069df755d224c8d2f27bc9246603b93131aa667e5e1f5e492
-size 111766
+oid sha256:910ef049b67560a279f8e0e702d74c1343f36c53d95ef84a32782c08c92a184a
+size 111866
