+++ apps/settings/signals.py
@@ -1,0 +1,3 @@
+from django.dispatch import Signal
+
+category_setting_updated = Signal(providing_args=('category', 'serializer'))
